Séparation de l'api et utilisateur
